Service and Controller with User, Log, Word table

+++ src/main/java/com/weblearnel/repository/UserRepository.java
@@ -1,9 +1,8 @@
 package com.weblearnel.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.weblearnel.model.User;
 
-public interface UserRepository extends JpaRepository<User, Integer>{
+public interface UserRepository extends JpaRepository<User, Long>{
     
 }
